pyproject.toml
[project]
name = "nutritracker"
version = "0.1.0"
description = ""
authors = [
    {name = "amitsingh-1811",email = "humhiamithai@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy[asyncio] (>=2.0.43,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "fastapi (>=0.117.1,<0.118.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "passlib[bcrypt] (==1.7.4)",
    "pydantic[email] (>=2.11.9,<3.0.0)",
    "bcrypt (==4.0.1)",
    "redis[async] (>=6.4.0,<7.0.0)",
    "aiosmtplib (>=4.0.2,<5.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)"
]

[tool.poetry]
packages = [{include = "nutritracker", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
